Set OutputStream.delegate to nil in HTTPBodyOutputStreamBridge.deinit
When running the cancellation tests in a loop, very occasionally there
would be a crash with the following backtrace:

```
```

This seems to indicate that the output stream is trying to access its
delegate. However, when running with debug logging enabled I can see
that the delegate has already been deinitialized.

This is likely a result of the delegate itself owning the stream and
setting the stream delegate to `self`, which IIUC is an established
pattern. This presents a race in teardown.

This patch sets the output stream delegate to `nil` in the delegate
`deinit`.

With this patch, the failing test passes when run an order of magnitude
more times than were required to reliably reproduce the crash without
the patch.