Tolerate both CancellationError and URLError in cancellation tests
When cancelling a Swift concurrency task during a streaming request
then the error returned might be `URLError` with `.cancelled` code or
`CancellationError`. The tests tried to be smart and expect just one of
these depending on which stage of the request we were at, but there are
still some races, and this test fails very rarely because
a `CancellationError` was thrown instead of a `URLError`.

This patch updates the test to tolerate both kinds of error at this
stage of the request.

Now the tests do not fail when run repeatedly.